refactor(passkey): extract passkey logic into dedicated trait and interface
- Extract passkey-related ORM fields and methods from IdentityTrait into new IdentityPasskeysTrait
- Add HasPasskeys interface under Traits namespace for type-safe passkey contract
- Remove passkey methods from Identity interface (getPasskeyUserHandle, setPasskeyUserHandle)
- Update FancyAdminExtension imports to reference new IdentityPasskeysTrait and HasPasskeys
